Refactor cluster distribution related optimizer rules into 7 files
