+++ arangod/Aql/OptimizerRules.h
@@ -108,16 +108,6 @@ void clusterPushSubqueryToDBServer(Optimizer* opt,
 void scatterInClusterRule(Optimizer*, std::unique_ptr<ExecutionPlan>,
                           OptimizerRule const&);
 
-/// @brief distribute operations in cluster - send each incoming row to every
-/// remote client precisely once. This happens in queries like:
-///
-///   FOR x IN coll1 [SOME FILTER/CALCULATION STATEMENTS] REMOVE f(x) in coll2
-///
-/// where coll2 is sharded by _key, but not if it is sharded by anything else.
-/// The collections coll1 and coll2 do not have to be distinct for this.
-void distributeInClusterRule(Optimizer*, std::unique_ptr<ExecutionPlan>,
-                             OptimizerRule const&);
-
 #ifdef USE_ENTERPRISE
 void distributeOffsetInfoToClusterRule(aql::Optimizer* opt,
                                        std::unique_ptr<aql::ExecutionPlan> plan,
@@ -146,56 +136,6 @@ void removeDistributeNodesRule(Optimizer*, std::unique_ptr<ExecutionPlan>,
 void smartJoinsRule(Optimizer*, std::unique_ptr<ExecutionPlan>,
                     OptimizerRule const&);
 #endif
-
-/// @brief try to restrict fragments to a single shard if possible
-void restrictToSingleShardRule(Optimizer*, std::unique_ptr<ExecutionPlan>,
-                               OptimizerRule const&);
-
-/// @brief move collect to the DB servers in cluster
-void collectInClusterRule(Optimizer*, std::unique_ptr<ExecutionPlan>,
-                          OptimizerRule const&);
-
-void distributeFilterCalcToClusterRule(Optimizer*,
-                                       std::unique_ptr<ExecutionPlan>,
-                                       OptimizerRule const&);
-
-void distributeSortToClusterRule(Optimizer*, std::unique_ptr<ExecutionPlan>,
-                                 OptimizerRule const&);
-
-/// @brief try to get rid of a RemoteNode->ScatterNode combination which has
-/// only a SingletonNode and possibly some CalculationNodes as dependencies
-void removeUnnecessaryRemoteScatterRule(Optimizer*,
-                                        std::unique_ptr<ExecutionPlan>,
-                                        OptimizerRule const&);
-
-/// @brief this rule removes Remote-Gather-Scatter/Distribute-Remote nodes from
-/// plans arising from queries of the form:
-///
-///   FOR x IN coll [SOME FILTER/CALCULATION STATEMENTS] REMOVE x IN coll
-///
-/// when coll is any collection sharded by any attributes, and
-///
-///   FOR x IN coll [SOME FILTER/CALCULATION STATEMENTS] REMOVE x._key IN coll
-///
-/// when the coll is sharded by _key.
-///
-/// This rule dues not work for queries like:
-///
-///    FOR x IN coll [SOME FILTER/CALCULATION STATEMENTS] REMOVE y IN coll
-///
-/// [different variable names], or
-///
-///   FOR x IN coll1 [SOME FILTER/CALCULATION STATEMENTS] REMOVE x in coll2
-///
-///  when coll1 and coll2 are not the same, or
-///
-///   FOR x IN coll [SOME FILTER/CALCULATION STATEMENTS] REMOVE f(x) in coll
-///
-///  where f is some function.
-///
-void undistributeRemoveAfterEnumCollRule(Optimizer*,
-                                         std::unique_ptr<ExecutionPlan>,
-                                         OptimizerRule const&);
 
 /// @brief remove $OLD and $NEW variables from data-modification statements
 /// if not required
