+++ arangod/Aql/OptimizerRulesFeature.cpp
@@ -51,6 +51,13 @@
 #include "Aql/Optimizer/Rule/InlineSubqueries.h"
 #include "Aql/Optimizer/Rule/OptimizeSubqueries.h"
 #include "Aql/Optimizer/Rule/SpliceSubqueries.h"
+#include "Aql/Optimizer/Rule/DistributeInCluster.h"
+#include "Aql/Optimizer/Rule/CollectInCluster.h"
+#include "Aql/Optimizer/Rule/DistributeFilterCalcToCluster.h"
+#include "Aql/Optimizer/Rule/DistributeSortToCluster.h"
+#include "Aql/Optimizer/Rule/RemoveUnnecessaryRemoteScatter.h"
+#include "Aql/Optimizer/Rule/RestrictToSingleShard.h"
+#include "Aql/Optimizer/Rule/UndistributeRemoveAfterEnumColl.h"
 #include "Aql/OptimizerRules.h"
 #include "Aql/OptimizerRulesOptionsProvider.h"
 #include "Basics/Exceptions.h"
